INT-7183: Use two columns for forms
Create two cols for forms, required and advanced and add elements to
each. Various small CSS tweaks to make it work better or look
slightly nicer.

On iPad and smaller float the buttons at the bottom so they are
always onscreen.

Downgrade cancel button to a link, so it's less visible.