INT-21029 theme_snap: Added My Courses to the snap_blocks rendering funnel
The my/courses.php page has it's very sui generis way of rendering blocks.
After discussions with the team, it was decided that Snap should closely
follow the behavior from Core. To achieve this, the blocks from the block drawer
should only be rendered inside the block drawer (they used to render at the
bottom of the page).

The blocks from the dashboard template, do still continue to render at the
bottom of the page, same as the Core observed behavior.